kvserver: move helper to where it's used
Move newOnDiskEngine helper to replica_sst_snapshot_storage_test.go, the only
place where it's used.

Release note: None
